refactor: stretch — isolate remaining KV, retire api mirrors, internalize Ask leaves
Create dedicated Cloudflare KV namespaces for contact messages, AI response
cache, and AI feedback. Stop generating public/api search JSON and redirect
legacy URLs to /search/*. Move Ask leaf hooks under hooks/internal/ so the
public barrel stays five composed surfaces.
