--- public/api/blog.json
@@ -1,85 +1,0 @@
-[
-  {
-    "slug": "ethics-in-the-ai-age-semcacfe",
-    "title": "Ethics in the AI Age: Speaking at SEMCACFE",
-    "description": "Notes from co-presenting Ethics in the AI Age with UHY at SEMCACFE — why professional skepticism still beats confident AI output, and what changes when models start taking actions.",
-    "publishedAt": "2026-07-21",
-    "tags": [
-      "Speaking",
-      "AI Ethics",
-      "Agentic AI",
-      "Governance"
-    ],
-    "author": "Blake Oxford",
-    "draft": false,
-    "featured": true
-  },
-  {
-    "slug": "ces-2026-ai-has-left-the-screen",
-    "title": "CES 2026: AI Has Left the Screen — and That Changes Everything",
-    "description": "At CES 2026, AI felt less like a chat demo and more like a system embedded in real workflows — from factories and cars to robotics and simulation.",
-    "publishedAt": "2026-01-12",
-    "tags": [
-      "Blog",
-      "CES",
-      "AI",
-      "Robotics",
-      "Voice AI",
-      "Generative AI",
-      "NVIDIA"
-    ],
-    "author": "Blake Oxford",
-    "draft": false,
-    "featured": false
-  },
-  {
-    "slug": "building-my-own-local-llm-stack",
-    "title": "Building My Own Local LLM Stack (Because Sometimes You Shouldn't Ask ChatGPT About a Rash)",
-    "description": "How I built a secure, self-healing local AI stack using Ollama, Docker, and Open WebUI on a Mac Mini — a private, practical ChatGPT alternative.",
-    "publishedAt": "2025-10-25",
-    "tags": [
-      "local LLM",
-      "Ollama",
-      "Open WebUI",
-      "DevOps",
-      "privacy",
-      "AI infrastructure",
-      "macOS"
-    ],
-    "author": "Blake Oxford",
-    "draft": false,
-    "featured": true
-  },
-  {
-    "slug": "ai-statistics-future-decision-making",
-    "title": "AI, Statistics, and the Future of Decision-Making: What Our Research Reveals",
-    "description": "From our GPT-4 study to GPT-5's reasoning models — exploring RAG, CAG, MCP, and the evolution toward augmented, auditable analytics systems that amplify human capability.",
-    "publishedAt": "2025-10-14",
-    "tags": [
-      "AI",
-      "Machine Learning",
-      "Statistics",
-      "Research",
-      "Leadership",
-      "GPT-4",
-      "Data Science"
-    ],
-    "author": "Blake Oxford",
-    "draft": false,
-    "featured": true
-  },
-  {
-    "slug": "combating-legal-ai-hallucinations",
-    "title": "Combating Legal AI Hallucinations: How courtlistener-mcp Enhances Trustworthy AI-Assisted Legal Research",
-    "description": "Introducing courtlistener-mcp, an open-source MCP server that combats AI hallucinations in legal research by grounding responses in real court data from CourtListener.",
-    "publishedAt": "2025-07-28",
-    "tags": [
-      "AI",
-      "Legal Tech",
-      "Open Source"
-    ],
-    "author": "Blake Oxford",
-    "draft": false,
-    "featured": true
-  }
-]
