--- public/api/projects.json
@@ -1,236 +1,0 @@
-[
-  {
-    "slug": "fanalyx-deterministic-finance-platform",
-    "title": "Fanalyx Deterministic Finance Platform",
-    "description": "",
-    "publishedAt": "2025-03-01",
-    "tags": [
-      "Financial Modeling",
-      "Deterministic AI",
-      "Cloudflare Workers",
-      "TypeScript",
-      "Zod",
-      "MCP",
-      "Personal Finance"
-    ],
-    "draft": false,
-    "technologies": [
-      "Financial Modeling",
-      "Deterministic AI",
-      "Cloudflare Workers",
-      "TypeScript",
-      "Zod",
-      "MCP",
-      "Personal Finance"
-    ],
-    "image": "/assets/images/og-image.jpg",
-    "categories": [
-      "AI & Data",
-      "Finance",
-      "Edge Computing"
-    ],
-    "featured": true
-  },
-  {
-    "slug": "microsoft-fabric",
-    "title": "Microsoft Fabric – Operational Intelligence & Workflow Automation",
-    "description": "",
-    "publishedAt": "2024-02-20",
-    "tags": [
-      "Microsoft Fabric",
-      "Power BI",
-      "Power Platform",
-      "Process Automation",
-      "EOS",
-      "Data Engineering",
-      "Leadership"
-    ],
-    "draft": false,
-    "technologies": [
-      "Microsoft Fabric",
-      "Power BI",
-      "Power Platform",
-      "Process Automation",
-      "EOS",
-      "Data Engineering",
-      "Leadership"
-    ],
-    "image": "/assets/projects/operational-and-workflow-automation.png",
-    "categories": [
-      "Automation",
-      "AI & Data",
-      "Healthcare IT"
-    ],
-    "featured": true
-  },
-  {
-    "slug": "ferment-app",
-    "title": "Ferment App – Mobile Recipe Management",
-    "description": "",
-    "publishedAt": "2024-01-15",
-    "tags": [
-      "Swift",
-      "SwiftUI",
-      "Firebase",
-      "SwiftData",
-      "Mobile Development",
-      "Fermentation"
-    ],
-    "draft": false,
-    "technologies": [
-      "Swift",
-      "SwiftUI",
-      "Firebase",
-      "SwiftData",
-      "Mobile Development",
-      "Fermentation"
-    ],
-    "image": "/assets/projects/ferment-app-design.png",
-    "categories": [
-      "Mobile & Product",
-      "Fermentation"
-    ],
-    "featured": true
-  },
-  {
-    "slug": "llm-note-coaching",
-    "title": "OpenAI-Powered Documentation Quality Feedback System",
-    "description": "",
-    "publishedAt": "2023-11-01",
-    "tags": [
-      "OpenAI",
-      "Natural Language Processing",
-      "Healthcare IT",
-      "Compliance",
-      "Blue Cross Blue Shield",
-      "Documentation Quality",
-      "Python"
-    ],
-    "draft": false,
-    "technologies": [
-      "OpenAI",
-      "Natural Language Processing",
-      "Healthcare IT",
-      "Compliance",
-      "Blue Cross Blue Shield",
-      "Documentation Quality",
-      "Python"
-    ],
-    "image": "/assets/projects/openai-automated-audit.png",
-    "categories": [
-      "AI & Data",
-      "Healthcare IT"
-    ],
-    "featured": false
-  },
-  {
-    "slug": "adp-workforcenow",
-    "title": "ADP Workforce Now Implementation",
-    "description": "",
-    "publishedAt": "2021-01-01",
-    "tags": [
-      "HCM",
-      "Recruiting",
-      "ADP Workforce Now",
-      "Sage Intacct",
-      "PowerBI"
-    ],
-    "draft": false,
-    "technologies": [
-      "HCM",
-      "Recruiting",
-      "ADP Workforce Now",
-      "Sage Intacct",
-      "PowerBI"
-    ],
-    "image": "/assets/projects/adp-automation.png",
-    "categories": [
-      "Operations",
-      "Automation"
-    ],
-    "featured": false
-  },
-  {
-    "slug": "google-workspace-migration",
-    "title": "Google Workspace → Microsoft 365 Migration & Endpoint Management",
-    "description": "",
-    "publishedAt": "2020-01-15",
-    "tags": [
-      "Microsoft 365",
-      "Google Workspace",
-      "Endpoint Manager",
-      "Intune",
-      "MDM",
-      "COVID-19 Response"
-    ],
-    "draft": false,
-    "technologies": [
-      "Microsoft 365",
-      "Google Workspace",
-      "Endpoint Manager",
-      "Intune",
-      "MDM",
-      "COVID-19 Response"
-    ],
-    "image": "/assets/projects/google-to-microsoft.png",
-    "categories": [
-      "Operations",
-      "COVID-19 Response"
-    ],
-    "featured": false
-  },
-  {
-    "slug": "bank-projections-modeling",
-    "title": "Bank Projections and Financial Modeling",
-    "description": "",
-    "publishedAt": "2020-01-01",
-    "tags": [
-      "Financial Modeling",
-      "Bank Projections",
-      "Commercial Real Estate",
-      "Loan",
-      "Financial Analysis"
-    ],
-    "draft": false,
-    "technologies": [
-      "Financial Modeling",
-      "Bank Projections",
-      "Commercial Real Estate",
-      "Loan",
-      "Financial Analysis"
-    ],
-    "image": "/assets/projects/bank-projections.png",
-    "categories": [
-      "Financial Intelligence",
-      "Operations"
-    ],
-    "featured": false
-  },
-  {
-    "slug": "advancedmd-implementation",
-    "title": "AdvancedMD Implementation & Evolution",
-    "description": "",
-    "publishedAt": "2017-12-01",
-    "tags": [
-      "EHR",
-      "Digital Transformation",
-      "Automation",
-      "Healthcare IT",
-      "SQL"
-    ],
-    "draft": false,
-    "technologies": [
-      "EHR",
-      "Digital Transformation",
-      "Automation",
-      "Healthcare IT",
-      "SQL"
-    ],
-    "image": "/assets/projects/advancedMD-project.png",
-    "categories": [
-      "Healthcare IT",
-      "Operations"
-    ],
-    "featured": false
-  }
-]
