+++ src/features/chat/hooks/README.md
@@ -1,0 +1,11 @@
+# Ask chat hooks
+
+Public surface (import from `@/features/chat/hooks`):
+
+- `useAIChatController`
+- `useChatInteraction`
+- `useChatPersistence`
+- `useChatStreaming`
+- `useChatSession`
+
+Everything under `internal/` is private to those composed surfaces. Do not import leaf hooks from outside this folder.
