+++ src/features/chat/hooks/internal/useAiChatBridge.ts
