+++ src/features/chat/hooks/internal/useChatLifecycle.ts
