+++ src/features/chat/hooks/internal/useComputedValues.ts
