+++ src/features/chat/hooks/internal/useQueryManagement.ts
