+++ src/features/chat/hooks/internal/useScrollManagement.ts
