+++ tests/contracts/baselines/projects.baseline.json
@@ -1,5 +1,5 @@
 {
-  "source": "public/api/projects.json",
+  "source": "public/search/projects.json",
   "schemaVersion": 1,
   "entries": [
     {
